Sampling is no more a global event, but is machine related. Moving "Sampling.handle" to DraggableEvent.handle, using the new Event interface. Now only NonWorkingDay is a global event (handled by the handle method).
Plan to use the same method to handle local event that requires to shift
other ones.

Working on sampling.jsp. Working on program.jsp to move sampling events
addition and deletion to this calendar.

Is now required the ability to redraw every event after shift (redraw or
refresh, find out the most elegant solution or if the elegant solution
is too hard, find a way to make this work).